#5560 rewrite defaultFocus
