+++ js/notebook/src/plot/plotUtils.js
@@ -282,91 +282,6 @@ define([
       labelHeight : 12,
       tooltipWidth : 10
     },
-    getDefaultFocus : function(model) {
-      var yAxisData = [], yAxisRData = [];
-      for (var i = 0; i < model.data.length; i++) {
-        var item = model.data[i];
-        if(plotUtils.useYAxisR(model, item)){
-          yAxisRData.push(item);
-        } else {
-          yAxisData.push(item);
-        }
-      }
-
-      var ret = this.getDataRange(yAxisData);
-      var retR = this.getDataRange(yAxisRData);
-      var range = ret.datarange;
-      var rangeR = retR.datarange;
-      var margin = model.margin;
-
-      if(ret.visibleItem === 0) { // for empty plot, focus needs to be adjusted
-        range.xl = model.xAxis.getPercent(range.xl);
-        range.xr = model.xAxis.getPercent(range.xr);
-        range.yl = model.yAxis.getPercent(range.yl);
-        range.yr = model.yAxis.getPercent(range.yr);
-        if (model.yAxisR) {
-          rangeR.yl = model.yAxisR.getPercent(rangeR.yl);
-          rangeR.yr = model.yAxisR.getPercent(rangeR.yr);
-        }
-      }
-      var focus = {
-        xl : model.userFocus.xl,
-        xr : model.userFocus.xr,
-        yl : model.userFocus.yl,
-        yr : model.userFocus.yr,
-        yl_r : model.yAxisR !== null ? model.userFocus.yl_r : model.userFocus.yl,
-        yr_r : model.yAxisR !== null ? model.userFocus.yr_r : model.userFocus.yr
-      };
-
-      if (focus.xl == null) {
-        focus.xl = this.minus(range.xl, this.mult(range.xspan, margin.left));
-      }
-      if (focus.xr == null) {
-        focus.xr = this.plus(range.xr, this.mult(range.xspan, margin.right));
-      }
-      if (focus.xl instanceof Big) {
-        focus.xl = parseFloat(focus.xl.toString());
-      }
-      if (focus.xr instanceof Big) {
-        focus.xr = parseFloat(focus.xr.toString());
-      }
-
-      if (focus.yl == null) {
-        if (model.yIncludeZero === true) {
-          var yl = model.vrange.yspan * range.yl + model.vrange.yl;
-          if(yl > 0){
-            range.yl = (0 - model.vrange.yl) / model.vrange.yspan;
-            range.yspan = range.yr - range.yl;
-          }
-        }
-        focus.yl = range.yl - range.yspan * margin.bottom;
-      }
-      if (focus.yr == null) {
-        focus.yr = range.yr + range.yspan * margin.top;
-      }
-
-      if (focus.yl_r == null) {
-        if (model.yIncludeZero === true && model.vrangeR) {
-          var yl_r = model.vrangeR.yspan * rangeR.yl + model.vrangeR.yl;
-          if(yl_r > 0){
-            rangeR.yl = (0 - model.vrangeR.yl) / model.vrangeR.yspan;
-            rangeR.yspan = rangeR.yr - rangeR.yl;
-          }
-        }
-        focus.yl_r = rangeR.yl - rangeR.yspan * (_.isNumber(margin.bottom_r) ? margin.bottom_r : 0);
-      }
-      if (focus.yr_r == null) {
-        focus.yr_r = rangeR.yr + rangeR.yspan * (_.isNumber(margin.top_r) ? margin.top_r : 0);
-      }
-
-      focus.xspan = focus.xr - focus.xl;
-      focus.yspan = focus.yr - focus.yl;
-      focus.yspan_r = focus.yr_r - focus.yl_r;
-      var result = {};
-      result.defaultFocus = focus;
-      _.extend(result, _.omit(ret, "datarange"));
-      return result;
-    },
     safeWidth: function(e){
       return bkHelper.isChrome ? this.getComputedStyle(e, 'width') : e.width();
     },
